requirements.txt
beautifulsoup4==4.12.3
certifi==2024.7.4
chardet==5.2.0
idna==3.7
requests==2.32.3
urllib3==2.2.2
ruff==0.5.0
python-dotenv==1.0.1